OS-6702-1: Datasource disconnect survey fixes
